+++ Team01/Game/Project/Enemy.h
@@ -34,15 +34,15 @@ public:
 		float dir_rotation;
 		int bullet_gap;
 		int bullet_setgap;
-		int hp_max;
+		float hp_max;
 		std::string texture_path;
-		InitParam(Mof::CVector2 pos, int move, int pinch_move, float ratio, float time, int column, int amount, int set, int reflect, float cdir, float dirrange, float dirrotat, int bgap, int bsetgap, int hp, std::string tex_path) :
+		InitParam(Mof::CVector2 pos, int move, int pinch_move, float ratio, float time, int column, int amount, int set, int reflect, float cdir, float dirrange, float dirrotat, int bgap, int bsetgap, float hp, std::string tex_path) :
 			position(pos), move_type(move), move_type_on_pinch(pinch_move), pinch_hp_ratio(ratio), spawn_time(time), bullet_column(column), bullet_amount(amount), amount_set(set), reflect_count(reflect), central_dir(cdir), dir_range(dirrange), dir_rotation(dirrotat), bullet_gap(bgap), bullet_setgap(bsetgap), hp_max(hp), texture_path(tex_path) {
 		}
 	};
 private:
-	int m_HP;
-	int m_MaxHP;
+	float m_HP;
+	float m_MaxHP;
 	Vector2 m_InitPos;
 	Vector2 m_Pos;
 	Vector2 m_Move;
@@ -115,7 +115,12 @@ public:
 	bool IsShow(void) const { return m_bDrow; }
 
 	void Initialize(const InitParam& param);
-	void Initialize(Vector2 pos, int move_type, int pinch_move, float ratio, int column, int amount, int set, int reflect, float cdir, float dirrange, float dirrotat, int bgap, int bsetgap, int hp);
+	void Initialize(
+		Vector2 pos, int move_type, int pinch_move, 
+		float ratio, int column, int amount, 
+		int set, int reflect, float cdir, 
+		float dirrange, float dirrotat, 
+		int bgap, int bsetgap, float hp);
 	void SetTexture(Mof::CTexture* ptr);
 	void SetTarget(Mof::CVector2 pos);
 	void Update(bool end);
@@ -128,7 +133,7 @@ public:
 	/// </summary>
 	/// <param name=""></param>
 	/// <returns></returns>
-	bool Damage(int damage_value);
+	bool Damage(float damage_value);
 	int GetFastBulletNo() { return m_FastBulletNo; }
 	void SetFastBulletNo(int no) { m_FastBulletNo = no; }
 	Mof::CRectangle GetCollisionRectangle(void) const;
